MonsterTick moved off a queued ActionChain to the new Tick mechanism
