+++ Source/ACE.Server/WorldObjects/Creature_Tick.cs
@@ -8,7 +8,21 @@ namespace ACE.Server.WorldObjects
             foreach (var wo in EquippedObjects.Values)
                 wo.Tick(lastTickDuration, currentTimeTick);
 
+            Monster_Tick(lastTickDuration, currentTimeTick);
+
             base.Tick(lastTickDuration, currentTimeTick);
+        }
+
+        /// <summary>
+        /// Called every ~5 seconds for Creatures
+        /// </summary>
+        public override void HeartBeat()
+        {
+            VitalTick();
+
+            // item enchantment ticks?
+
+            base.HeartBeat();
         }
     }
 }
